Changes to be committed: 	modified:   MOD_Up_x_Uq.f90 	modified:   exp_T200_JCP.dat 	modified:   exp_T200_REASSGND.dat
Change the input file that contains the states, energies and intensities.
The change was needed becaus of the way that the Intel Compiler (ifort)
reads the unformatted data. Now the input file contains line with
quantum numbers, energies and intensities separated by comas ',' and with
NO spaces.
This system requieres a relative big Nq, so the dimension is too big and
the Intel Compiler helps us to save time.